drf/requirements.txt
Django~=5.1.6
mysqlclient
djangorestframework>=3.14.0